diffText/requirements.txt
aim==3.19.2
aim-ui==3.19.2
aimrecords==0.0.7
aimrocks==0.4.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
attrs==23.2.0
base58==2.0.1
boto3==1.34.85
botocore==1.34.85
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
datasets==2.18.0
dill==0.3.8
einops==0.7.0
fastapi==0.110.1
filelock==3.13.4
frozenlist==1.4.1
fsspec==2024.2.0
greenlet==3.0.3
h11==0.14.0
huggingface-hub==0.22.2
hydra-core==1.3.2
idna==3.7
Jinja2==3.1.3
jmespath==1.0.1
Mako==1.3.3
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opencv-python-headless==4.7.0.72
oxenai
packaging==24.0
pandas==2.0.1
pillow==10.3.0
polars==0.17.11
psutil==5.9.8
pyarrow==15.0.2
pyarrow-hotfix==0.6
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
RestrictedPython==7.1
s3transfer==0.10.1
safetensors==0.4.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
sympy==1.12
tokenizers==0.15.2
toml==0.10.2
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.65.0
transformers==4.39.3
triton==2.2.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
websockets==12.0
xxhash==3.4.1
yarl==1.9.4
